+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT RELEASE
 
 * Removes the `params` from the `Address->verify()` method since it's non-static and unused
+* Add `columns` and `additional_columns` support to `Report->create()`
 
 ## v5.0.0 (2022-02-09)